Revert "Refactor users package (#1129)"
This reverts commit ceaeeb1abd297f9be090ac93492812f61b935438.
