+++ examples/resources/genesyscloud_flow/inboundcall_flow_example.yaml
@@ -1,5 +1,5 @@
 inboundCall:
-  name: Terraform Flow Test-e678ea2e-2f37-4376-ab79-d70e7c7359b7
+  name: Terraform Emergency Test Flow 7fa2e8a9-b1d9-4d78-93b8-05b4465314a7
   defaultLanguage: en-us
   startUpRef: ./menus/menu[mainMenu]
   initialGreeting:
