+++ examples/resources/genesyscloud_flow/inboundcall_flow_example2.yaml
@@ -1,6 +1,6 @@
 inboundEmail:
-    name: Terraform Flow Test-dabfb12a-6e7e-4e8c-ab21-7f8549959c48
-    division: New Home1
+    name: Terraform Flow Test-d0c40522-88e4-4b66-aa48-f237afb819e0
+    division: New Home
     startUpRef: "/inboundEmail/states/state[Initial State_10]"
     defaultLanguage: en-us
     supportedLanguages:
