+++ examples/resources/genesyscloud_flow/inboundcall_flow_example3.yaml
@@ -1,5 +1,5 @@
 inboundCall:
-  name: Terraform Flow Test-fa4dd927-bf74-4394-9b20-c3560a4b5f7f
+  name: Terraform Flow Test-e602c7ba-318e-4373-99bb-bdf69dcc7600
   defaultLanguage: en-us
   startUpRef: ./menus/menu[mainMenu]
   initialGreeting:
