+++ genesyscloud/group/genesyscloud_group_init_test.go
@@ -1,11 +1,10 @@
 package group
 
 import (
-	"sync"
-	"terraform-provider-genesyscloud/genesyscloud/user"
-	"testing"
-
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"sync"
+	"terraform-provider-genesyscloud/genesyscloud"
+	"testing"
 )
 
 var providerDataSources map[string]*schema.Resource
@@ -24,7 +23,7 @@ func (r *registerTestInstance) registerTestResources() {
 	defer r.resourceMapMutex.Unlock()
 
 	providerResources[resourceName] = ResourceGroup()
-	providerResources["genesyscloud_user"] = user.ResourceUser()
+	providerResources["genesyscloud_user"] = genesyscloud.ResourceUser()
 
 }
