+++ genesyscloud/integration/genesyscloud_integration_init_test.go
@@ -3,9 +3,10 @@ package integration
 import (
 	"sync"
 	"terraform-provider-genesyscloud/genesyscloud/group"
-	integrationCred "terraform-provider-genesyscloud/genesyscloud/integration_credential"
-	"terraform-provider-genesyscloud/genesyscloud/user"
 	"testing"
+
+	gcloud "terraform-provider-genesyscloud/genesyscloud"
+	integrationCred "terraform-provider-genesyscloud/genesyscloud/integration_credential"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
@@ -36,7 +37,7 @@ func (r *registerTestInstance) registerTestResources() {
 	providerResources["genesyscloud_integration"] = ResourceIntegration()
 	providerResources["genesyscloud_group"] = group.ResourceGroup()
 	providerResources["genesyscloud_integration_credential"] = integrationCred.ResourceIntegrationCredential()
-	providerResources["genesyscloud_user"] = user.ResourceUser()
+	providerResources["genesyscloud_user"] = gcloud.ResourceUser()
 }
 
 // registerTestDataSources registers all data sources used in the tests.
