+++ genesyscloud/resource_genesyscloud_user_test.go
@@ -1,4 +1,4 @@
-package user
+package genesyscloud
 
 import (
 	"context"
@@ -6,12 +6,11 @@ import (
 	"log"
 	"strconv"
 	"strings"
-	"terraform-provider-genesyscloud/genesyscloud"
 	"terraform-provider-genesyscloud/genesyscloud/provider"
 	routinglanguage "terraform-provider-genesyscloud/genesyscloud/routing_language"
-	routingSkill "terraform-provider-genesyscloud/genesyscloud/routing_skill"
 	routingUtilization "terraform-provider-genesyscloud/genesyscloud/routing_utilization"
 	routingUtilizationLabel "terraform-provider-genesyscloud/genesyscloud/routing_utilization_label"
+	routingSkill "terraform-provider-genesyscloud/genesyscloud/routing_skill"
 
 	extensionPool "terraform-provider-genesyscloud/genesyscloud/telephony_providers_edges_extension_pool"
 	"terraform-provider-genesyscloud/genesyscloud/util"
@@ -66,17 +65,17 @@ func TestAccResourceUserBasic(t *testing.T) {
 					"",             // No certs
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "name", userName1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "state", stateActive),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "title", title1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "department", department1),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "password.%"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "manager", ""),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "acd_auto_answer", "false"),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "profile_skills.%"),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "certifications.%"),
-					provider.TestDefaultHomeDivision(resourceName+"."+userResource1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "name", userName1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "state", stateActive),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "title", title1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "department", department1),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "password.%"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "manager", ""),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "acd_auto_answer", "false"),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "profile_skills.%"),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "certifications.%"),
+					provider.TestDefaultHomeDivision("genesyscloud_user."+userResource1),
 				),
 			},
 			{
@@ -94,16 +93,16 @@ func TestAccResourceUserBasic(t *testing.T) {
 					strconv.Quote(cert1),
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "name", userName2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "state", stateInactive),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "title", title2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "department", department2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "manager", ""),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "acd_auto_answer", "true"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "profile_skills.0", profileSkill1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "certifications.0", cert1),
-					provider.TestDefaultHomeDivision(resourceName+"."+userResource1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "name", userName2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "state", stateInactive),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "title", title2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "department", department2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "manager", ""),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "acd_auto_answer", "true"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "profile_skills.0", profileSkill1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "certifications.0", cert1),
+					provider.TestDefaultHomeDivision("genesyscloud_user."+userResource1),
 				),
 			},
 			{
@@ -126,18 +125,18 @@ func TestAccResourceUserBasic(t *testing.T) {
 					util.NullValue, // Active
 					strconv.Quote(title1),
 					strconv.Quote(department1),
-					resourceName+"."+userResource1+".id",
+					"genesyscloud_user."+userResource1+".id",
 					util.TrueValue, // AcdAutoAnswer
 					strconv.Quote(profileSkill1),
 					strconv.Quote(cert1),
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource2, "email", email3),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource2, "name", userName1),
-					resource.TestCheckResourceAttrPair(resourceName+"."+userResource2, "manager", resourceName+"."+userResource1, "id"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "manager", ""),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "profile_skills.0", profileSkill2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "certifications.0", cert2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource2, "email", email3),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource2, "name", userName1),
+					resource.TestCheckResourceAttrPair("genesyscloud_user."+userResource2, "manager", "genesyscloud_user."+userResource1, "id"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "manager", ""),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "profile_skills.0", profileSkill2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "certifications.0", cert2),
 				),
 			},
 			{
@@ -155,16 +154,16 @@ func TestAccResourceUserBasic(t *testing.T) {
 					"",              // No certs
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource2, "email", email3),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource2, "name", userName1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource2, "manager", ""),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource2, "profile_skills.%"),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource2, "certifications.%"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource2, "email", email3),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource2, "name", userName1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource2, "manager", ""),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource2, "profile_skills.%"),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource2, "certifications.%"),
 				),
 			},
 			{
 				// Import/Read
-				ResourceName:      resourceName + "." + userResource2,
+				ResourceName:      "genesyscloud_user." + userResource2,
 				ImportState:       true,
 				ImportStateVerify: true,
 			},
@@ -174,12 +173,12 @@ func TestAccResourceUserBasic(t *testing.T) {
 }
 
 func generateUserWithCustomAttrs(resourceID string, email string, name string, attrs ...string) string {
-	return fmt.Sprintf(`resource "%s" "%s" {
+	return fmt.Sprintf(`resource "genesyscloud_user" "%s" {
 		email = "%s"
 		name = "%s"
 		%s
 	}
-	`, resourceName, resourceID, email, name, strings.Join(attrs, "\n"))
+	`, resourceID, email, name, strings.Join(attrs, "\n"))
 }
 
 func TestAccResourceUserAddresses(t *testing.T) {
@@ -201,8 +200,8 @@ func TestAccResourceUserAddresses(t *testing.T) {
 		addrTypeWork              = "WORK"
 		addrTypeHome              = "HOME"
 		extensionPoolResource1    = "test-extensionpool1" + uuid.NewString()
-		extensionPoolStartNumber1 = "8000"
-		extensionPoolEndNumber1   = "9000"
+		extensionPoolStartNumber1 = "1000"
+		extensionPoolEndNumber1   = "2000"
 	)
 
 	extensionPoolResource := extensionPool.ExtensionPoolStruct{
@@ -251,7 +250,7 @@ func TestAccResourceUserAddresses(t *testing.T) {
 			},
 			{
 				// Import/Read
-				ResourceName:      resourceName + "." + addrUserResource1,
+				ResourceName:      "genesyscloud_user." + addrUserResource1,
 				ImportState:       true,
 				ImportStateVerify: true,
 			},
@@ -431,12 +430,12 @@ func TestAccResourceUserPhone(t *testing.T) {
 					),
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+addrUserResource1, "email", addrEmail1),
-					resource.TestCheckResourceAttr(resourceName+"."+addrUserResource1, "name", addrUserName),
-					resource.TestCheckResourceAttr(resourceName+"."+addrUserResource1, "addresses.0.phone_numbers.0.number", addrPhone2),
-					resource.TestCheckResourceAttr(resourceName+"."+addrUserResource1, "addresses.0.phone_numbers.0.extension", addrExt1),
-					resource.TestCheckResourceAttr(resourceName+"."+addrUserResource1, "addresses.0.phone_numbers.0.media_type", phoneMediaType),
-					resource.TestCheckResourceAttr(resourceName+"."+addrUserResource1, "addresses.0.phone_numbers.0.type", addrTypeWork),
+					resource.TestCheckResourceAttr("genesyscloud_user."+addrUserResource1, "email", addrEmail1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+addrUserResource1, "name", addrUserName),
+					resource.TestCheckResourceAttr("genesyscloud_user."+addrUserResource1, "addresses.0.phone_numbers.0.number", addrPhone2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+addrUserResource1, "addresses.0.phone_numbers.0.extension", addrExt1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+addrUserResource1, "addresses.0.phone_numbers.0.media_type", phoneMediaType),
+					resource.TestCheckResourceAttr("genesyscloud_user."+addrUserResource1, "addresses.0.phone_numbers.0.type", addrTypeWork),
 				),
 			},
 		},
@@ -471,7 +470,7 @@ func TestAccResourceUserSkills(t *testing.T) {
 					generateUserRoutingSkill("genesyscloud_routing_skill."+skillResource1+".id", proficiency1),
 				) + routingSkill.GenerateRoutingSkillResource(skillResource1, skillName1),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserSkill(resourceName+"."+userResource1, "genesyscloud_routing_skill."+skillResource1, proficiency1),
+					validateUserSkill("genesyscloud_user."+userResource1, "genesyscloud_routing_skill."+skillResource1, proficiency1),
 				),
 			},
 			{
@@ -490,8 +489,8 @@ func TestAccResourceUserSkills(t *testing.T) {
 					skillName2,
 				),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserSkill(resourceName+"."+userResource1, "genesyscloud_routing_skill."+skillResource1, proficiency1),
-					validateUserSkill(resourceName+"."+userResource1, "genesyscloud_routing_skill."+skillResource2, proficiency2),
+					validateUserSkill("genesyscloud_user."+userResource1, "genesyscloud_routing_skill."+skillResource1, proficiency1),
+					validateUserSkill("genesyscloud_user."+userResource1, "genesyscloud_routing_skill."+skillResource2, proficiency2),
 				),
 			},
 			{
@@ -506,7 +505,7 @@ func TestAccResourceUserSkills(t *testing.T) {
 					skillName2,
 				),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserSkill(resourceName+"."+userResource1, "genesyscloud_routing_skill."+skillResource2, proficiency1),
+					validateUserSkill("genesyscloud_user."+userResource1, "genesyscloud_routing_skill."+skillResource2, proficiency1),
 				),
 			},
 			{
@@ -557,7 +556,7 @@ func TestAccResourceUserLanguages(t *testing.T) {
 					generateUserRoutingLang("genesyscloud_routing_language."+langResource1+".id", proficiency1),
 				) + routinglanguage.GenerateRoutingLanguageResource(langResource1, langName1),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserLanguage(resourceName+"."+userResource1, "genesyscloud_routing_language."+langResource1, proficiency1),
+					validateUserLanguage("genesyscloud_user."+userResource1, "genesyscloud_routing_language."+langResource1, proficiency1),
 				),
 			},
 			{
@@ -576,8 +575,8 @@ func TestAccResourceUserLanguages(t *testing.T) {
 					langName2,
 				),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserLanguage(resourceName+"."+userResource1, "genesyscloud_routing_language."+langResource1, proficiency1),
-					validateUserLanguage(resourceName+"."+userResource1, "genesyscloud_routing_language."+langResource2, proficiency2),
+					validateUserLanguage("genesyscloud_user."+userResource1, "genesyscloud_routing_language."+langResource1, proficiency1),
+					validateUserLanguage("genesyscloud_user."+userResource1, "genesyscloud_routing_language."+langResource2, proficiency2),
 				),
 			},
 			{
@@ -592,7 +591,7 @@ func TestAccResourceUserLanguages(t *testing.T) {
 					langName2,
 				),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserLanguage(resourceName+"."+userResource1, "genesyscloud_routing_language."+langResource2, proficiency1),
+					validateUserLanguage("genesyscloud_user."+userResource1, "genesyscloud_routing_language."+langResource2, proficiency1),
 				),
 			},
 			{
@@ -604,7 +603,7 @@ func TestAccResourceUserLanguages(t *testing.T) {
 					"routing_languages = []",
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_languages.%"),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_languages.%"),
 					func(s *terraform.State) error {
 						time.Sleep(30 * time.Second) // Wait for 30 seconds for proper deletion
 						return nil
@@ -644,11 +643,11 @@ func TestAccResourceUserLocations(t *testing.T) {
 						"genesyscloud_location."+locResource1+".id",
 						strconv.Quote(locNotes1),
 					),
-				) + genesyscloud.GenerateLocationResourceBasic(locResource1, locName1),
+				) + GenerateLocationResourceBasic(locResource1, locName1),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email),
-					resource.TestCheckResourceAttrPair(resourceName+"."+userResource1, "locations.0.location_id", "genesyscloud_location."+locResource1, "id"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "locations.0.notes", locNotes1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email),
+					resource.TestCheckResourceAttrPair("genesyscloud_user."+userResource1, "locations.0.location_id", "genesyscloud_location."+locResource1, "id"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "locations.0.notes", locNotes1),
 				),
 			},
 			{
@@ -661,11 +660,11 @@ func TestAccResourceUserLocations(t *testing.T) {
 						"genesyscloud_location."+locResource2+".id",
 						strconv.Quote(locNotes2),
 					),
-				) + genesyscloud.GenerateLocationResourceBasic(locResource2, locName2),
+				) + GenerateLocationResourceBasic(locResource2, locName2),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email),
-					resource.TestCheckResourceAttrPair(resourceName+"."+userResource1, "locations.0.location_id", "genesyscloud_location."+locResource2, "id"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "locations.0.notes", locNotes2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email),
+					resource.TestCheckResourceAttrPair("genesyscloud_user."+userResource1, "locations.0.location_id", "genesyscloud_location."+locResource2, "id"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "locations.0.notes", locNotes2),
 				),
 			},
 		},
@@ -707,12 +706,12 @@ func TestAccResourceUserEmployerInfo(t *testing.T) {
 					),
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "name", userName),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.official_name", offName1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.employee_id", ""),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.employee_type", ""),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.date_hire", ""),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "name", userName),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.official_name", offName1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.employee_id", ""),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.employee_type", ""),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.date_hire", ""),
 				),
 			},
 			{
@@ -729,12 +728,12 @@ func TestAccResourceUserEmployerInfo(t *testing.T) {
 					),
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "name", userName),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.official_name", ""),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.employee_id", empID1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.employee_type", empTypeFull),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.date_hire", hireDate1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "name", userName),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.official_name", ""),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.employee_id", empID1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.employee_type", empTypeFull),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.date_hire", hireDate1),
 				),
 			},
 			{
@@ -751,12 +750,12 @@ func TestAccResourceUserEmployerInfo(t *testing.T) {
 					),
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "name", userName),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.official_name", offName2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.employee_id", empID2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.employee_type", empTypePart),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "employer_info.0.date_hire", hireDate2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "name", userName),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.official_name", offName2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.employee_id", empID2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.employee_type", empTypePart),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "employer_info.0.date_hire", hireDate2),
 				),
 			},
 			{
@@ -768,9 +767,9 @@ func TestAccResourceUserEmployerInfo(t *testing.T) {
 					"employer_info = []",
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "name", userName),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "employer_info.%"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "name", userName),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "employer_info.%"),
 				),
 			},
 		},
@@ -810,22 +809,22 @@ func TestAccResourceUserroutingUtilBasic(t *testing.T) {
 					),
 				),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserUtilizationLevel(resourceName+"."+userResource1, "Agent"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.FalseValue),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.interruptible_media_types.%"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.FalseValue),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types.%"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.FalseValue),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types.%"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.FalseValue),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.interruptible_media_types.%"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.FalseValue),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.interruptible_media_types.%"),
+					validateUserUtilizationLevel("genesyscloud_user."+userResource1, "Agent"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.FalseValue),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.interruptible_media_types.%"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.FalseValue),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types.%"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.FalseValue),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types.%"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.FalseValue),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.interruptible_media_types.%"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.FalseValue),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.interruptible_media_types.%"),
 				),
 			},
 			{
@@ -843,22 +842,22 @@ func TestAccResourceUserroutingUtilBasic(t *testing.T) {
 					),
 				),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserUtilizationLevel(resourceName+"."+userResource1, "Agent"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.call.0.interruptible_media_types", utilTypeEmail),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.email.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.message.0.interruptible_media_types", utilTypeCall),
+					validateUserUtilizationLevel("genesyscloud_user."+userResource1, "Agent"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.interruptible_media_types", utilTypeEmail),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.interruptible_media_types", utilTypeCall),
 				),
 			},
 			{
@@ -876,22 +875,22 @@ func TestAccResourceUserroutingUtilBasic(t *testing.T) {
 					),
 				),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserUtilizationLevel(resourceName+"."+userResource1, "Agent"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity0),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.call.0.interruptible_media_types", utilTypeEmail),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity0),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity0),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity0),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.email.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity0),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.message.0.interruptible_media_types", utilTypeCall),
+					validateUserUtilizationLevel("genesyscloud_user."+userResource1, "Agent"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity0),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.interruptible_media_types", utilTypeEmail),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity0),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity0),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity0),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity0),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.interruptible_media_types", utilTypeCall),
 				),
 			},
 			{
@@ -903,8 +902,8 @@ func TestAccResourceUserroutingUtilBasic(t *testing.T) {
 					"routing_utilization = []",
 				),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserUtilizationLevel(resourceName+"."+userResource1, "Organization"),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.%"),
+					validateUserUtilizationLevel("genesyscloud_user."+userResource1, "Organization"),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.%"),
 				),
 			},
 		},
@@ -956,26 +955,26 @@ func TestAccResourceUserroutingUtilWithLabels(t *testing.T) {
 						),
 					),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserUtilizationLevel(resourceName+"."+userResource1, "Agent"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.FalseValue),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.interruptible_media_types.%"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.FalseValue),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types.%"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.FalseValue),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types.%"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.FalseValue),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.interruptible_media_types.%"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.FalseValue),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.interruptible_media_types.%"),
-					resource.TestCheckResourceAttrSet(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.0.label_id"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.0.maximum_capacity", maxCapacity1),
-					resource.TestCheckResourceAttrSet(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.1.label_id"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.1.maximum_capacity", maxCapacity1),
+					validateUserUtilizationLevel("genesyscloud_user."+userResource1, "Agent"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.FalseValue),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.interruptible_media_types.%"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.FalseValue),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types.%"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.FalseValue),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types.%"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.FalseValue),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.interruptible_media_types.%"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.FalseValue),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.interruptible_media_types.%"),
+					resource.TestCheckResourceAttrSet("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.0.label_id"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.0.maximum_capacity", maxCapacity1),
+					resource.TestCheckResourceAttrSet("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.1.label_id"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.1.maximum_capacity", maxCapacity1),
 				),
 			},
 			{
@@ -998,26 +997,26 @@ func TestAccResourceUserroutingUtilWithLabels(t *testing.T) {
 						),
 					),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserUtilizationLevel(resourceName+"."+userResource1, "Agent"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.call.0.interruptible_media_types", utilTypeEmail),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.email.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.message.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttrSet(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.0.label_id"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.0.maximum_capacity", maxCapacity2),
-					resource.TestCheckResourceAttrSet(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.1.label_id"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.1.maximum_capacity", maxCapacity2),
+					validateUserUtilizationLevel("genesyscloud_user."+userResource1, "Agent"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.interruptible_media_types", utilTypeEmail),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttrSet("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.0.label_id"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.0.maximum_capacity", maxCapacity2),
+					resource.TestCheckResourceAttrSet("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.1.label_id"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.1.maximum_capacity", maxCapacity2),
 				),
 			},
 			{
@@ -1040,26 +1039,26 @@ func TestAccResourceUserroutingUtilWithLabels(t *testing.T) {
 						),
 					),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserUtilizationLevel(resourceName+"."+userResource1, "Agent"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity0),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.call.0.interruptible_media_types", utilTypeEmail),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity0),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity0),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity0),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.email.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity0),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.TrueValue),
-					util.ValidateStringInArray(resourceName+"."+userResource1, "routing_utilization.0.message.0.interruptible_media_types", utilTypeCall),
-					resource.TestCheckResourceAttrSet(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.0.label_id"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.0.maximum_capacity", maxCapacity0),
-					resource.TestCheckResourceAttrSet(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.1.label_id"),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "routing_utilization.0.label_utilizations.1.maximum_capacity", maxCapacity0),
+					validateUserUtilizationLevel("genesyscloud_user."+userResource1, "Agent"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.maximum_capacity", maxCapacity0),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.call.0.interruptible_media_types", utilTypeEmail),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.maximum_capacity", maxCapacity0),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.callback.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.maximum_capacity", maxCapacity0),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.chat.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.maximum_capacity", maxCapacity0),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.email.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.maximum_capacity", maxCapacity0),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.include_non_acd", util.TrueValue),
+					util.ValidateStringInArray("genesyscloud_user."+userResource1, "routing_utilization.0.message.0.interruptible_media_types", utilTypeCall),
+					resource.TestCheckResourceAttrSet("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.0.label_id"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.0.maximum_capacity", maxCapacity0),
+					resource.TestCheckResourceAttrSet("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.1.label_id"),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.0.label_utilizations.1.maximum_capacity", maxCapacity0),
 				),
 			},
 			{
@@ -1071,8 +1070,8 @@ func TestAccResourceUserroutingUtilWithLabels(t *testing.T) {
 					"routing_utilization = []",
 				),
 				Check: resource.ComposeTestCheckFunc(
-					validateUserUtilizationLevel(resourceName+"."+userResource1, "Organization"),
-					resource.TestCheckNoResourceAttr(resourceName+"."+userResource1, "routing_utilization.%"),
+					validateUserUtilizationLevel("genesyscloud_user."+userResource1, "Organization"),
+					resource.TestCheckNoResourceAttr("genesyscloud_user."+userResource1, "routing_utilization.%"),
 				),
 			},
 		},
@@ -1101,8 +1100,8 @@ func TestAccResourceUserRestore(t *testing.T) {
 					userName1,
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "name", userName1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "name", userName1),
 				),
 			},
 			{
@@ -1122,8 +1121,8 @@ func TestAccResourceUserRestore(t *testing.T) {
 					userName2,
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "name", userName2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "name", userName2),
 				),
 			},
 		},
@@ -1153,8 +1152,8 @@ func TestAccResourceUserCreateWhenDestroyed(t *testing.T) {
 					userName1,
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "name", userName1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "name", userName1),
 				),
 			},
 			{
@@ -1174,9 +1173,9 @@ func TestAccResourceUserCreateWhenDestroyed(t *testing.T) {
 					userName2,
 				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "email", email1),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "name", userName2),
-					resource.TestCheckResourceAttr(resourceName+"."+userResource1, "state", stateActive),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "email", email1),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "name", userName2),
+					resource.TestCheckResourceAttr("genesyscloud_user."+userResource1, "state", stateActive),
 				),
 			},
 		},
@@ -1189,7 +1188,7 @@ func testVerifyUsersDestroyed(state *terraform.State) error {
 
 	diagErr := util.WithRetries(context.Background(), 20*time.Second, func() *retry.RetryError {
 		for _, rs := range state.RootModule().Resources {
-			if rs.Type != resourceName {
+			if rs.Type != "genesyscloud_user" {
 				continue
 			}
 			err := checkUserDeleted(rs.Primary.ID)(state)
@@ -1202,10 +1201,10 @@ func testVerifyUsersDestroyed(state *terraform.State) error {
 				if util.IsStatus404(resp) {
 					continue
 				}
-				return retry.NonRetryableError(util.BuildWithRetriesApiDiagnosticError(resourceName, fmt.Sprintf("Unexpected error: %s", err), resp))
+				return retry.NonRetryableError(util.BuildWithRetriesApiDiagnosticError("genesyscloud_user", fmt.Sprintf("Unexpected error: %s", err), resp))
 			}
 
-			return retry.RetryableError(util.BuildWithRetriesApiDiagnosticError(resourceName, fmt.Sprintf("User (%s) still exists", rs.Primary.ID), resp))
+			return retry.RetryableError(util.BuildWithRetriesApiDiagnosticError("genesyscloud_user", fmt.Sprintf("User (%s) still exists", rs.Primary.ID), resp))
 		}
 		return nil
 	})
