+++ genesyscloud/routing_queue/resource_genesyscloud_routing_queue_test.go
@@ -11,11 +11,10 @@ import (
 	"terraform-provider-genesyscloud/genesyscloud/architect_user_prompt"
 	"terraform-provider-genesyscloud/genesyscloud/group"
 	"terraform-provider-genesyscloud/genesyscloud/provider"
-	routingSkill "terraform-provider-genesyscloud/genesyscloud/routing_skill"
 	routingSkillGroup "terraform-provider-genesyscloud/genesyscloud/routing_skill_group"
-	"terraform-provider-genesyscloud/genesyscloud/user"
 	"terraform-provider-genesyscloud/genesyscloud/util"
 	featureToggles "terraform-provider-genesyscloud/genesyscloud/util/feature_toggles"
+	routingSkill "terraform-provider-genesyscloud/genesyscloud/routing_skill"
 	"terraform-provider-genesyscloud/genesyscloud/util/testrunner"
 	"testing"
 	"time"
@@ -835,7 +834,7 @@ func TestAccResourceRoutingQueueSkillgroupMembers(t *testing.T) {
 					skillResourceId,
 					skillName,
 				) + skillGroupConfig + user2Config +
-					user.GenerateBasicUserResource(
+					genesyscloud.GenerateBasicUserResource(
 						user1ResourceId,
 						user1Email,
 						user1Name,
@@ -880,11 +879,11 @@ func TestAccResourceRoutingQueueMembers(t *testing.T) {
 		Steps: []resource.TestStep{
 			{
 				// Create
-				Config: user.GenerateBasicUserResource(
+				Config: genesyscloud.GenerateBasicUserResource(
 					queueMemberResource1,
 					queueMemberEmail1,
 					queueMemberName1,
-				) + user.GenerateBasicUserResource(
+				) + genesyscloud.GenerateBasicUserResource(
 					queueMemberResource2,
 					queueMemberEmail2,
 					queueMemberName2,
@@ -903,11 +902,11 @@ func TestAccResourceRoutingQueueMembers(t *testing.T) {
 					time.Sleep(30 * time.Second)
 				},
 				// Update with another queue member and modify rings
-				Config: user.GenerateBasicUserResource(
+				Config: genesyscloud.GenerateBasicUserResource(
 					queueMemberResource1,
 					queueMemberEmail1,
 					queueMemberName1,
-				) + user.GenerateBasicUserResource(
+				) + genesyscloud.GenerateBasicUserResource(
 					queueMemberResource2,
 					queueMemberEmail2,
 					queueMemberName2,
@@ -927,7 +926,7 @@ func TestAccResourceRoutingQueueMembers(t *testing.T) {
 			},
 			{
 				// Remove a queue member
-				Config: user.GenerateBasicUserResource(
+				Config: genesyscloud.GenerateBasicUserResource(
 					queueMemberResource2,
 					queueMemberEmail2,
 					queueMemberName2,
