+++ genesyscloud/routing_queue_conditional_group_routing/genesyscloud_routing_queue_conditional_group_routing_init_test.go
@@ -1,14 +1,13 @@
 package routing_queue_conditional_group_routing
 
 import (
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"sync"
+	gcloud "terraform-provider-genesyscloud/genesyscloud"
 	"terraform-provider-genesyscloud/genesyscloud/group"
 	routingQueue "terraform-provider-genesyscloud/genesyscloud/routing_queue"
 	routingSkillGroup "terraform-provider-genesyscloud/genesyscloud/routing_skill_group"
-	"terraform-provider-genesyscloud/genesyscloud/user"
 	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
 /*
@@ -31,7 +30,7 @@ func (r *registerTestInstance) registerTestResources() {
 	providerResources[resourceName] = ResourceRoutingQueueConditionalGroupRouting()
 	providerResources["genesyscloud_routing_queue"] = routingQueue.ResourceRoutingQueue()
 	providerResources["genesyscloud_routing_skill_group"] = routingSkillGroup.ResourceRoutingSkillGroup()
-	providerResources["genesyscloud_user"] = user.ResourceUser()
+	providerResources["genesyscloud_user"] = gcloud.ResourceUser()
 	providerResources["genesyscloud_group"] = group.ResourceGroup()
 }
