+++ genesyscloud/routing_skill_group/genesyscloud_routing_skill_group_init_test.go
@@ -3,9 +3,8 @@ package routing_skill_group
 import (
 	"sync"
 	"terraform-provider-genesyscloud/genesyscloud"
-	routingSkill "terraform-provider-genesyscloud/genesyscloud/routing_skill"
-	"terraform-provider-genesyscloud/genesyscloud/user"
 	"testing"
+	routingSkill "terraform-provider-genesyscloud/genesyscloud/routing_skill"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
@@ -25,7 +24,7 @@ func (r *registerTestInstance) registerTestResources() {
 
 	providerResources[resourceName] = ResourceRoutingSkillGroup()
 	providerResources["genesyscloud_auth_division"] = genesyscloud.ResourceAuthDivision()
-	providerResources["genesyscloud_user"] = user.ResourceUser()
+	providerResources["genesyscloud_user"] = genesyscloud.ResourceUser()
 	providerResources["genesyscloud_routing_skill"] = routingSkill.ResourceRoutingSkill()
 }
