+++ genesyscloud/routing_utilization_label/data_source_genesyscloud_routing_utilization_label_test.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/mypurecloud/platform-client-sdk-go/v133/platformclientv2"
 )
 
 func TestAccDataSourceRoutingUtilizationLabel(t *testing.T) {
@@ -23,7 +22,7 @@ func TestAccDataSourceRoutingUtilizationLabel(t *testing.T) {
 		ProviderFactories: provider.GetProviderFactories(providerResources, providerDataSources),
 		Steps: []resource.TestStep{
 			{
-				Config: generateRoutingUtilizationLabelResource(
+				Config: GenerateRoutingUtilizationLabelResource(
 					resourceName,
 					labelName,
 					"",
@@ -48,27 +47,4 @@ func generateRoutingUtilizationLabelDataSource(
         depends_on=[%s]
 	}
 	`, resourceID, name, dependsOnResource)
-}
-
-func checkIfLabelsAreEnabled() error { // remove once the feature is globally enabled
-	api := platformclientv2.NewRoutingApiWithConfig(sdkConfig)
-	_, resp, _ := api.GetRoutingUtilizationLabels(100, 1, "", "")
-	if resp.StatusCode == 501 {
-		return fmt.Errorf("feature is not yet implemented in this org.")
-	}
-	return nil
-}
-
-func generateRoutingUtilizationLabelResource(resourceID string, name string, dependsOnResource string) string {
-	dependsOn := ""
-
-	if dependsOnResource != "" {
-		dependsOn = fmt.Sprintf("depends_on=[genesyscloud_routing_utilization_label.%s]", dependsOnResource)
-	}
-
-	return fmt.Sprintf(`resource "genesyscloud_routing_utilization_label" "%s" {
-		name = "%s"
-		%s
-	}
-	`, resourceID, name, dependsOn)
 }
