+++ genesyscloud/station/data_source_genesyscloud_station_test.go
@@ -3,11 +3,11 @@ package station
 import (
 	"fmt"
 	"strconv"
+	gcloud "terraform-provider-genesyscloud/genesyscloud"
 	"terraform-provider-genesyscloud/genesyscloud/provider"
 	edgePhone "terraform-provider-genesyscloud/genesyscloud/telephony_providers_edges_phone"
 	phoneBaseSettings "terraform-provider-genesyscloud/genesyscloud/telephony_providers_edges_phonebasesettings"
 	edgeSite "terraform-provider-genesyscloud/genesyscloud/telephony_providers_edges_site"
-	"terraform-provider-genesyscloud/genesyscloud/user"
 	"terraform-provider-genesyscloud/genesyscloud/util"
 	"testing"
 
@@ -40,7 +40,7 @@ func TestAccDataSourceStation(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	config := user.GenerateUserResource(
+	config := gcloud.GenerateUserResource(
 		userRes1,
 		userEmail1,
 		userName1,
