+++ genesyscloud/tfexporter/resource_genesyscloud_tf_export_test.go
@@ -22,7 +22,6 @@ import (
 	resourceExporter "terraform-provider-genesyscloud/genesyscloud/resource_exporter"
 	routingQueue "terraform-provider-genesyscloud/genesyscloud/routing_queue"
 	telephonyProvidersEdgesSite "terraform-provider-genesyscloud/genesyscloud/telephony_providers_edges_site"
-	"terraform-provider-genesyscloud/genesyscloud/user"
 	"terraform-provider-genesyscloud/genesyscloud/util"
 	"testing"
 	"time"
@@ -607,7 +606,7 @@ func TestAccResourceTfExportByName(t *testing.T) {
 		Steps: []resource.TestStep{
 			{
 				// Generate a user and export it
-				Config: user.GenerateBasicUserResource(
+				Config: gcloud.GenerateBasicUserResource(
 					userResource1,
 					userEmail1,
 					userName1,
@@ -615,7 +614,7 @@ func TestAccResourceTfExportByName(t *testing.T) {
 			},
 			{
 				// Generate a user and export it
-				Config: user.GenerateBasicUserResource(
+				Config: gcloud.GenerateBasicUserResource(
 					userResource1,
 					userEmail1,
 					userName1,
@@ -637,7 +636,7 @@ func TestAccResourceTfExportByName(t *testing.T) {
 			},
 			{
 				// Generate a queue as well and export it
-				Config: user.GenerateBasicUserResource(
+				Config: gcloud.GenerateBasicUserResource(
 					userResource1,
 					userEmail1,
 					userName1,
@@ -679,7 +678,7 @@ func TestAccResourceTfExportByName(t *testing.T) {
 			},
 			{
 				// Export all trunk base settings as well
-				Config: user.GenerateBasicUserResource(
+				Config: gcloud.GenerateBasicUserResource(
 					userResource1,
 					userEmail1,
 					userName1,
@@ -730,11 +729,11 @@ func TestAccResourceTfExportByName(t *testing.T) {
 			},
 			{
 				// Export all trunk base settings as well
-				Config: user.GenerateBasicUserResource(
+				Config: gcloud.GenerateBasicUserResource(
 					userResource1,
 					userEmail1,
 					userName1,
-				) + user.GenerateBasicUserResource(
+				) + gcloud.GenerateBasicUserResource(
 					userResource2,
 					userEmail2,
 					userName2,
@@ -2812,7 +2811,7 @@ func buildQueueResources(queueExports []QueueExport) string {
 func buildUserResources(userExports []UserExport) string {
 	userResourceDefinitions := ""
 	for _, userExport := range userExports {
-		userResourceDefinitions = userResourceDefinitions + user.GenerateBasicUserResource(
+		userResourceDefinitions = userResourceDefinitions + gcloud.GenerateBasicUserResource(
 			userExport.ResourceName,
 			userExport.Email,
 			userExport.Name,
