+++ genesyscloud/user_roles/resource_genesyscloud_user_roles_test.go
@@ -7,7 +7,6 @@ import (
 	"terraform-provider-genesyscloud/genesyscloud"
 	authRole "terraform-provider-genesyscloud/genesyscloud/auth_role"
 	"terraform-provider-genesyscloud/genesyscloud/provider"
-	"terraform-provider-genesyscloud/genesyscloud/user"
 	"terraform-provider-genesyscloud/genesyscloud/util"
 	"terraform-provider-genesyscloud/genesyscloud/util/lists"
 	"testing"
@@ -43,7 +42,7 @@ func TestAccResourceUserRolesMembership(t *testing.T) {
 			{
 				// Create user with 1 role in default division
 				// Also add employee role reference as new user's automatically get this role
-				Config: "data \"genesyscloud_auth_division_home\" \"home\" {}\n" + user.GenerateBasicUserResource(
+				Config: "data \"genesyscloud_auth_division_home\" \"home\" {}\n" + genesyscloud.GenerateBasicUserResource(
 					userResource1,
 					email1,
 					userName1,
@@ -70,7 +69,7 @@ func TestAccResourceUserRolesMembership(t *testing.T) {
 			},
 			{
 				// Create another role and division and add to the user
-				Config: "data \"genesyscloud_auth_division_home\" \"home\" {}\n" + user.GenerateBasicUserResource(
+				Config: "data \"genesyscloud_auth_division_home\" \"home\" {}\n" + genesyscloud.GenerateBasicUserResource(
 					userResource1,
 					email1,
 					userName1,
@@ -99,7 +98,7 @@ func TestAccResourceUserRolesMembership(t *testing.T) {
 			},
 			{
 				// Remove a role from the user and modify division
-				Config: user.GenerateBasicUserResource(
+				Config: genesyscloud.GenerateBasicUserResource(
 					userResource1,
 					email1,
 					userName1,
@@ -118,7 +117,7 @@ func TestAccResourceUserRolesMembership(t *testing.T) {
 			},
 			{
 				// Remove all roles from the user
-				Config: user.GenerateBasicUserResource(
+				Config: genesyscloud.GenerateBasicUserResource(
 					userResource1,
 					email1,
 					userName1,
